cleaning go install script

+++ scripts/install-go.sh
@@ -1,14 +1,19 @@
 #!/bin/bash
 set -e
+./check-root.sh
 
-## Note that this will install golang but you'll need to add it to your
-## path if it's not already in your profile with `export PATH=$PATH:/usr/local/go/bin`
-## ref: https://golang.org/doc/install
+source ./add-to-path.sh
+GO_VERSION=1.17.3
 
-v=1.16
-os=linux-amd64
-tarname=go${v}.${os}.tar.gz
-wget https://golang.org/dl/${tarname}
-tar -C /usr/local -xzf ${tarname}
-rm -f ${tarname} 
-
+if go version ; then
+  echo golang already exists
+  exit 0
+else
+  echo Installing go...
+  ## ref: https://golang.org/doc/install
+  os=linux-amd64
+  tarname=go${GO_VERSION}.${os}.tar.gz
+  wget https://golang.org/dl/${tarname}
+  tar -C /usr/local -xzf ${tarname}
+  rm -f ${tarname} 
+fi
